# src/RuleService/src/main/java/com/orchestrator/rules/model/Member.java
package com.orchestrator.rules.model;

public class Member {
    private String membershipId;
    private String status;
    private String email;

    public String getMembershipId() { return membershipId; }
    public void setMembershipId(String membershipId) { this.membershipId = membershipId; }

    public String getStatus() { return status; }
    public void setStatus(String status) { this.status = status; }

    public String getEmail() { return email; }
    public void setEmail(String email) { this.email = email; }
}