# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ContextProviderService",
    "version": "1.0"
  },
  "paths": {
    "/api/Consistency/wait": {
      "post": {
        "tags": [
          "Consistency"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsistencyWaitRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsistencyWaitRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ConsistencyWaitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ConsistencyResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsistencyResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsistencyResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Context/resolve": {
      "get": {
        "tags": [
          "Context"
        ],
        "parameters": [
          {
            "name": "membershipId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "profile",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FactPayload"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FactPayload"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FactPayload"
                }
              }
            }
          }
        }
      }
    },
    "/": {
      "get": {
        "tags": [
          "ContextProviderService"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/Members": {
      "get": {
        "tags": [
          "Members"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Member"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Member"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Member"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Members"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMemberRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMemberRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMemberRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Member"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Member"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Member"
                }
              }
            }
          }
        }
      }
    },
    "/api/Members/{id}": {
      "get": {
        "tags": [
          "Members"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Member"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Member"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Member"
                }
              }
            }
          }
        }
      }
    },
    "/api/Members/by-membership/{membershipId}": {
      "get": {
        "tags": [
          "Members"
        ],
        "parameters": [
          {
            "name": "membershipId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Member"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Member"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Member"
                }
              }
            }
          }
        }
      }
    },
    "/api/Members/{id}/status": {
      "put": {
        "tags": [
          "Members"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStatusRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStatusRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Members/{id}/profile": {
      "put": {
        "tags": [
          "Members"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Members/{id}/wallet": {
      "put": {
        "tags": [
          "Members"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWalletRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWalletRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWalletRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ComplianceFact": {
        "type": "object",
        "properties": {
          "riskLevel": {
            "type": "string",
            "nullable": true
          },
          "kycStatus": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ConsistencyResponse": {
        "type": "object",
        "properties": {
          "isConsistent": {
            "type": "boolean"
          },
          "currentPos": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ConsistencyWaitRequest": {
        "type": "object",
        "properties": {
          "entityType": {
            "type": "string",
            "nullable": true
          },
          "entityId": {
            "type": "string",
            "nullable": true
          },
          "requiredMinPos": {
            "type": "integer",
            "format": "int64"
          },
          "timeoutMs": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "CreateMemberRequest": {
        "type": "object",
        "properties": {
          "membershipId": {
            "type": "string",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/MemberStatus"
          },
          "risk_Level": {
            "$ref": "#/components/schemas/RiskLevel"
          },
          "kyC_Level": {
            "$ref": "#/components/schemas/KycLevel"
          },
          "walletStatus": {
            "$ref": "#/components/schemas/WalletStatus"
          },
          "gameStatus": {
            "$ref": "#/components/schemas/GameStatus"
          },
          "currency": {
            "$ref": "#/components/schemas/Currency"
          }
        },
        "additionalProperties": false
      },
      "Currency": {
        "enum": [
          "CNY",
          "THB",
          "VND"
        ],
        "type": "string"
      },
      "FactPayload": {
        "type": "object",
        "properties": {
          "member": {
            "$ref": "#/components/schemas/MemberFact"
          },
          "wallet": {
            "$ref": "#/components/schemas/WalletFact"
          },
          "compliance": {
            "$ref": "#/components/schemas/ComplianceFact"
          }
        },
        "additionalProperties": false
      },
      "GameStatus": {
        "enum": [
          "Unlocked",
          "Locked"
        ],
        "type": "string"
      },
      "KycLevel": {
        "enum": [
          "Pending",
          "Verified",
          "Rejected",
          "UnderReview"
        ],
        "type": "string"
      },
      "Member": {
        "required": [
          "email",
          "email_Verified",
          "firstName",
          "kyC_Level",
          "lastName",
          "membershipId",
          "phone",
          "phone_Verified",
          "risk_Level",
          "status"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "membershipId": {
            "minLength": 1,
            "type": "string"
          },
          "firstName": {
            "minLength": 1,
            "type": "string"
          },
          "lastName": {
            "minLength": 1,
            "type": "string"
          },
          "email": {
            "minLength": 1,
            "type": "string",
            "format": "email"
          },
          "phone": {
            "minLength": 1,
            "type": "string",
            "format": "tel"
          },
          "status": {
            "$ref": "#/components/schemas/MemberStatus"
          },
          "risk_Level": {
            "$ref": "#/components/schemas/RiskLevel"
          },
          "kyC_Level": {
            "$ref": "#/components/schemas/KycLevel"
          },
          "email_Verified": {
            "type": "boolean"
          },
          "phone_Verified": {
            "type": "boolean"
          },
          "walletStatus": {
            "$ref": "#/components/schemas/WalletStatus"
          },
          "gameStatus": {
            "$ref": "#/components/schemas/GameStatus"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastPasswordChangedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastSkipSecurityUpdateAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "MemberFact": {
        "type": "object",
        "properties": {
          "membershipId": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "risk_Level": {
            "type": "string",
            "nullable": true
          },
          "kyC_Level": {
            "type": "string",
            "nullable": true
          },
          "email_Verified": {
            "type": "boolean"
          },
          "phone_Verified": {
            "type": "boolean"
          },
          "gameStatus": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "MemberStatus": {
        "enum": [
          "Active",
          "Inactive",
          "Suspended",
          "Confiscated"
        ],
        "type": "string"
      },
      "RiskLevel": {
        "enum": [
          "Low",
          "Medium",
          "High"
        ],
        "type": "string"
      },
      "UpdateProfileRequest": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "email_Verified": {
            "type": "boolean"
          },
          "phone_Verified": {
            "type": "boolean"
          },
          "gameStatus": {
            "$ref": "#/components/schemas/GameStatus"
          },
          "risk_Level": {
            "$ref": "#/components/schemas/RiskLevel"
          },
          "kyC_Level": {
            "$ref": "#/components/schemas/KycLevel"
          },
          "walletStatus": {
            "$ref": "#/components/schemas/WalletStatus"
          }
        },
        "additionalProperties": false
      },
      "UpdateStatusRequest": {
        "type": "object",
        "properties": {
          "status": {
            "$ref": "#/components/schemas/MemberStatus"
          }
        },
        "additionalProperties": false
      },
      "UpdateWalletRequest": {
        "type": "object",
        "properties": {
          "walletStatus": {
            "$ref": "#/components/schemas/WalletStatus"
          }
        },
        "additionalProperties": false
      },
      "WalletFact": {
        "type": "object",
        "properties": {
          "balance": {
            "type": "number",
            "format": "double"
          },
          "currency": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "WalletStatus": {
        "enum": [
          "Unlocked",
          "Locked",
          "Frozen"
        ],
        "type": "string"
      }
    }
  }
}